Refactored basic navigation class
