+++ ros/launch/basic_navigation.launch
@@ -10,26 +10,14 @@
     <arg name="goal_topic" default="/move_base_simple/goal"/>
     <arg name="localisation_topic" default="/amcl_pose"/>
     <arg name="cmd_vel_topic" default="/load/cmd_vel"/>
-    <arg name="global_planner_service" default="/global_planner/make_plan"/>
-    <arg name="nav_feedback_topic" default="/route_navigation/feedback"/>
-
-    <!-- global planner will also launch global costmap -->
-    <node pkg="navfn" name="global_planner" type="navfn_node" output="screen" >
-        <rosparam command="load" ns="costmap"
-                  file="$(find basic_navigation)/ros/config/$(arg robot_name)/common_costmap_params.yaml"/>
-        <rosparam command="load" ns="costmap"
-                  file="$(find basic_navigation)/ros/config/$(arg robot_name)/global_costmap_params.yaml"/>
-    </node>
 
     <node pkg="basic_navigation" name="basic_navigation" type="basic_navigation_node" output="screen" >
         <rosparam command="load"
         file="$(find basic_navigation)/ros/config/$(arg robot_name)/basic_navigation_params.yaml" />
-        <param name="global_planner_service" type="string" value="$(arg global_planner_service)"/>
         <remap from="~laser" to="$(arg laser_topic)"/>
         <remap from="~odom" to="$(arg odom_topic)"/>
         <remap from="~goal" to="$(arg goal_topic)"/>
         <remap from="~localisation" to="$(arg localisation_topic)"/>
         <remap from="~cmd_vel" to="$(arg cmd_vel_topic)"/>
-        <remap from="~nav_feedback" to="$(arg nav_feedback_topic)"/>
     </node>
 </launch>
