+++ ros/scripts/basic_navigation_node
@@ -2,7 +2,7 @@
 
 import rospy
 import traceback
-from basic_navigation.basic_navigation import BasicNavigation
+from basic_navigation.basic_navigation_class import BasicNavigation
 
 if __name__ == "__main__":
     rospy.init_node('basic_navigation')
